Fix implementation of TourGuideNPCTests
The engine needs to be defined in the setUp() method to be able to be
called for all the tests. Create TourGuideNPCTest public method within
the class. Change name of testHi() to TourGuideDialogueTest(). Add in
new TourGuideNPC() to the setupZone function call otherwise the test
will always fail because the NPC has never been initialised.

Implements: tests for tour guide NPC

Co-authored-by: e66406ob <omar.badawi-2@student.manchester.ac.uk>