Create TourGuideNPC class add NPC to Ados city .xml file
Create the TourGuideNPC class and initialise all of the NPC's traits,
speech etc. This will cause the TourGuideNPCTests to not fail. Currently
doesn't have a walking path this is to be implemented later. This class
sets his appearance and dialogue. Add TourGuideNPC to the ados.xml file
so that he appears on the map, otherwise he will not be added to
ados.city.

Implements: TourGuideNPC creation and addition to the map
Co-authored-by: e66406ob <omar.badawi-2@student.manchester.ac.uk>